+++ CHANGELOG.md
@@ -11,10 +11,14 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## (3.14.0 Candidate)
 ### Added
 - Engine 
+  - Flynn._3DMeshText() renders text to a 3D mesh object.
   - Flynn._3DMesh now has .PreRotation
   - color alpha support
     - Flynn._3DMesh.init() now accepts alpha (defaults to 1.0)
     - ctx.vectoStart() now accepts alpha (defaults to 1.0)
+- Test Application
+  - Add text meshes to 3D demo panel.
+  - Add alpha color demos.
 
 ## 3.13.1 - 2019-01-11
 ### Fixed